+++ ydb/public/sdk/cpp/include/ydb-cpp-sdk/client/topic/control_plane.h
@@ -1,6 +1,7 @@
 #pragma once
 
 #include "codecs.h"
+#include "read_events.h"
 
 #include <ydb/public/sdk/cpp/include/ydb-cpp-sdk/client/scheme/scheme.h>
 
@@ -122,22 +123,6 @@ private:
     TDuration MaxReadTimeLag_;
     TDuration MaxWriteTimeLag_;
     TDuration MaxCommittedTimeLag_;
-};
-
-// Topic partition location
-class TPartitionLocation {
-public:
-    TPartitionLocation(const Ydb::Topic::PartitionLocation& partitionLocation);
-    TPartitionLocation(i32 nodeId, i64 generation);
-    int32_t GetNodeId() const;
-    int64_t GetGeneration() const;
-
-private:
-    // Node identificator.
-    int32_t NodeId_ = 1;
-
-    // Partition generation.
-    int64_t Generation_ = 2;
 };
 
 class TPartitionInfo {
