+++ ydb/public/sdk/cpp/src/client/topic/impl/read_session_impl.ipp
@@ -6,7 +6,7 @@
 
 #include "direct_reader.h"
 
-#include <ydb/public/sdk/cpp/include/ydb-cpp-sdk/client/topic/control_plane.h>
+// #include <ydb/public/sdk/cpp/include/ydb-cpp-sdk/client/topic/control_plane.h>
 
 #include <ydb/public/sdk/cpp/src/client/topic/common/log_lazy.h>
