+++ ydb/public/sdk/cpp/src/client/topic/ya.make
@@ -13,7 +13,7 @@ PEERDIR(
     ydb/public/sdk/cpp/src/client/proto
     ydb/public/sdk/cpp/src/client/driver
     ydb/public/sdk/cpp/src/client/table
-    
+
     ydb/public/api/grpc
     ydb/public/api/grpc/draft
     ydb/public/api/protos
